[test] Rename input pattern files
