+++ test/input_error_common.js
