+++ test/input_ok_common.js
